refactor(grpc-client): deduplicate tokenizer proto and merge bundle files
Introduce common.proto for shared GetTokenizerRequest/GetTokenizerChunk
messages that were copy-pasted across all three engine protos (sglang,
vllm, trtllm). This eliminates proto-level duplication by having each
engine import common.proto and reference the shared types.

What changed:
- grpc_client/proto/common.proto: new shared proto with GetTokenizerRequest
  and GetTokenizerChunk messages under smg.grpc.common package
- grpc_client/proto/{sglang_scheduler,vllm_engine,trtllm_service}.proto:
  removed duplicated message definitions, added import of common.proto,
  updated RPC signatures to reference smg.grpc.common types
- grpc_client/build.rs: two-pass compilation — first pass compiles
  common.proto (types only, no gRPC services), second pass compiles
  engine protos with extern_path mapping .smg.grpc.common to
  crate::common_proto
- grpc_client/src/lib.rs: added common_proto module with
  tonic::include_proto, added impl_get_tokenizer! macro that provides
  the shared get_tokenizer() implementation for all engine clients
- grpc_client/src/tokenizer_bundle.rs: merged from stream_bundle.rs
  and archive_ops.rs into a single cohesive module containing stream
  collection, SHA-256 validation, zip validation, and extraction
- grpc_client/src/{sglang_scheduler,vllm_engine,trtllm_service}.rs:
  replaced manual get_tokenizer() with impl_get_tokenizer!() macro,
  removed stream_bundle imports
- grpc_client/src/stream_bundle.rs: deleted (merged into tokenizer_bundle)
- grpc_client/src/archive_ops.rs: deleted (merged into tokenizer_bundle)
- model_gateway/src/routers/grpc/client.rs: updated imports from
  archive_ops to tokenizer_bundle
- model_gateway/src/core/steps/tokenizer_registration.rs: updated
  imports from archive_ops/stream_bundle to tokenizer_bundle

Why:
The original PR (#470) copy-pasted identical GetTokenizerRequest and
GetTokenizerChunk proto messages into each of the three engine protos,
and duplicated the Rust get_tokenizer() client implementation across
all three engine clients. The tokenizer bundle handling was also split
across two tightly-coupled files (stream_bundle.rs for streaming,
archive_ops.rs for extraction) that logically belong together.

How:
Used tonic_prost_build's extern_path to map the shared proto package
to a single Rust module. A macro eliminates the identical 12-line
get_tokenizer() method across all three engine clients. The two-pass
build ensures common types are compiled first, then referenced by
engine protos. Net result: -525 lines, 14 files touched, all 46
tests pass.
