+++ grpc_client/build.rs
@@ -1,13 +1,21 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Rebuild triggers
+    println!("cargo:rerun-if-changed=proto/common.proto");
     println!("cargo:rerun-if-changed=proto/sglang_scheduler.proto");
     println!("cargo:rerun-if-changed=proto/vllm_engine.proto");
     println!("cargo:rerun-if-changed=proto/trtllm_service.proto");
 
-    // Compile protobuf files
+    // Pass 1: compile shared message types (no gRPC service generation)
+    tonic_prost_build::configure()
+        .build_server(false)
+        .build_client(false)
+        .compile_protos(&["proto/common.proto"], &["proto"])?;
+
+    // Pass 2: compile engine protos, referencing common types via extern_path
     tonic_prost_build::configure()
         .build_server(true)
         .build_client(true)
+        .extern_path(".smg.grpc.common", "crate::common_proto")
         .type_attribute("GetModelInfoResponse", "#[derive(serde::Serialize)]")
         // vllm + trtllm ServerInfo have only primitive fields.
         // sglang's contains prost_types::{Struct,Timestamp} so it's handled separately.
