+++ grpc_client/proto/sglang_scheduler.proto
@@ -4,6 +4,7 @@ package sglang.grpc.scheduler;
 
 import "google/protobuf/timestamp.proto";
 import "google/protobuf/struct.proto";
+import "common.proto";
 
 // Service definition for SGLang scheduler communication
 // This protocol bridges the Rust router and Python scheduler
@@ -30,7 +31,7 @@ service SglangScheduler {
   rpc GetLoads(GetLoadsRequest) returns (GetLoadsResponse);
 
   // Get tokenizer artifacts for remote construction
-  rpc GetTokenizer(GetTokenizerRequest) returns (stream GetTokenizerChunk);
+  rpc GetTokenizer(smg.grpc.common.GetTokenizerRequest) returns (stream smg.grpc.common.GetTokenizerChunk);
 
 }
 
@@ -587,16 +588,4 @@ message AggregateMetrics {
   double avg_token_usage = 4;
   double avg_throughput = 5;
   double avg_utilization = 6;
-}
-
-
-// =====================
-// Tokenizer
-// =====================
-
-message GetTokenizerRequest {}
-
-message GetTokenizerChunk {
-  bytes data = 1;       // Raw zip bytes (concatenate all chunks)
-  string sha256 = 2;    // SHA-256 fingerprint; set on final chunk only, empty on previous
 }
