+++ grpc_client/proto/trtllm_service.proto
@@ -16,6 +16,8 @@
 syntax = "proto3";
 package trtllm;
 
+import "common.proto";
+
 // TensorRT-LLM gRPC Service
 //
 // This service provides high-performance inference for LLMs using TensorRT-LLM.
@@ -48,7 +50,7 @@ service TrtllmService {
   rpc GetServerInfo(GetServerInfoRequest) returns (GetServerInfoResponse);
 
   // Get tokenizer artifacts for remote construction
-  rpc GetTokenizer(GetTokenizerRequest) returns (stream GetTokenizerChunk);
+  rpc GetTokenizer(smg.grpc.common.GetTokenizerRequest) returns (stream smg.grpc.common.GetTokenizerChunk);
 }
 
 // ============================================================================
@@ -507,15 +509,4 @@ message GetServerInfoResponse {
   int32 pipeline_parallel_size = 4; // PP size
   int32 context_parallel_size = 5;  // CP size
   int32 world_size = 6;             // Total world size
-}
-
-// ============================================================================
-// Tokenizer
-// ============================================================================
-
-message GetTokenizerRequest {}
-
-message GetTokenizerChunk {
-  bytes data = 1;       // Raw zip bytes (concatenate all chunks)
-  string sha256 = 2;    // SHA-256 fingerprint; set on final chunk only, empty on previous
 }
