+++ grpc_client/proto/vllm_engine.proto
@@ -2,6 +2,8 @@ syntax = "proto3";
 
 package vllm.grpc.engine;
 
+import "common.proto";
+
 // Service definition for vLLM engine communication
 // This protocol is designed for efficient binary communication between
 // the Rust router and vLLM Python engine (AsyncLLM).
@@ -25,7 +27,7 @@ service VllmEngine {
   rpc GetServerInfo(GetServerInfoRequest) returns (GetServerInfoResponse);
 
   // Get tokenizer artifacts for remote construction
-  rpc GetTokenizer(GetTokenizerRequest) returns (stream GetTokenizerChunk);
+  rpc GetTokenizer(smg.grpc.common.GetTokenizerRequest) returns (stream smg.grpc.common.GetTokenizerChunk);
 }
 
 // =====================
@@ -256,15 +258,4 @@ message GetServerInfoResponse {
   // KV transfer config (for PD disaggregation)
   string kv_connector = 6;  // "MooncakeConnector", "NixlConnector", "" if not configured
   string kv_role = 7;       // "kv_producer", "kv_consumer", "kv_both", "" if not configured
-}
-
-// =====================
-// Tokenizer
-// =====================
-
-message GetTokenizerRequest {}
-
-message GetTokenizerChunk {
-  bytes data = 1;       // Raw zip bytes (concatenate all chunks)
-  string sha256 = 2;    // SHA-256 fingerprint; set on final chunk only, empty on previous
 }
