+++ grpc_client/src/lib.rs
@@ -3,9 +3,12 @@
 //! This crate provides gRPC client implementations for communicating with
 //! SGLang scheduler, vLLM engine, and TensorRT-LLM engine backends.
 
-pub mod archive_ops;
+pub mod common_proto {
+    #![allow(clippy::all, clippy::absolute_paths, unused_qualifications)]
+    tonic::include_proto!("smg.grpc.common");
+}
 pub mod sglang_scheduler;
-pub mod stream_bundle;
+pub mod tokenizer_bundle;
 pub mod trtllm_service;
 pub mod vllm_engine;
 
@@ -16,6 +19,34 @@ pub use sglang_scheduler::{proto as sglang_proto, SglangSchedulerClient};
 use tonic::metadata::MetadataMap;
 pub use trtllm_service::{proto as trtllm_proto, TrtllmServiceClient};
 pub use vllm_engine::{proto as vllm_proto, VllmEngineClient};
+
+/// Shared `get_tokenizer()` implementation for all engine clients.
+///
+/// Each engine's generated proto client has a `get_tokenizer` RPC method
+/// with identical signature (using common proto types). This macro provides
+/// the wrapper that calls `collect_bundle_from_rpc` with the standard
+/// timeout and chunk extraction.
+macro_rules! impl_get_tokenizer {
+    () => {
+        pub async fn get_tokenizer(
+            &self,
+        ) -> Result<
+            $crate::tokenizer_bundle::StreamBundle,
+            Box<dyn std::error::Error + Send + Sync>,
+        > {
+            use $crate::common_proto::GetTokenizerRequest;
+            let request = tonic::Request::new(GetTokenizerRequest {});
+            let mut client = self.client.clone();
+            $crate::tokenizer_bundle::collect_bundle_from_rpc(
+                client.get_tokenizer(request),
+                |chunk| (chunk.data, chunk.sha256),
+                std::time::Duration::from_secs(120),
+            )
+            .await
+        }
+    };
+}
+pub(crate) use impl_get_tokenizer;
 
 /// Trait for injecting trace context into gRPC metadata.
 ///
